Main File Function Offload
Attempted to move functions into a separate header file to streamline things, but a "multiple definition" error was met and could not be troubleshooted; troubleshooting efforts made understanding the code more difficult. Due_Test_Jig_CC_2 was added as a rewrite of the firmware, but more friendly to header file application.